backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
websockets==12.0
redis==5.0.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
httpx==0.25.2
aiosmtplib==3.0.1
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
